moving flock code outside SaveConfig
